Update skill calculations, again
Had to fix updates for skill that have racial bonuses but aren't supposed to be active for the race selected, but still updates the totals based on skill points added.

Also removes this from the todo list, two birds, one commit.

Update code to reset calculations when switch races, currently non-active race skill bonuses from background and talents are applying.
Do not regress code so values aren't saved and reapplied for active race.
Possible fix, when switching races, all *_mdr is reset to *_skill_mdr and then bonuses for the active race are applied from backgrounds and talents for that race.

So it would be

reset *_mdr to *_skill_mdr (so you don't overwrite someone adding skill points
apply bonuses based on checkboxes checked for backgrounds and talents for the active race only.